Moves all custom exceptions to a separate file.
