+++ src/exceptions.hpp
@@ -1,0 +1,16 @@
+#pragma once
+#include <exception>
+
+namespace src_exceptions {
+class BackwardBeforeForwardException : public std::exception {
+  virtual const char *what() const throw() {
+    return "backward was called before forward.";
+  }
+};
+
+class InvalidShapeException : public std::exception {
+  virtual const char *what() const throw() {
+    return "An invalid shape was provided.";
+  }
+};
+} // namespace src_exceptions
